requirements.txt
PySide6
Pillow
PyQtWebEngine
numpy

